FINAL UPDATE

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: ahartley <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 09:22:57 by ahartley          #+#    #+#             */
-/*   Updated: 2019/06/07 16:54:26 by ahartley         ###   ########.fr       */
+/*   Updated: 2019/06/11 15:26:06 by ahartley         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_atoi(const char *str)
 {
-	int i;
-	int sign;
-	int ans;
+	int						i;
+	int						sign;
+	unsigned long long int	ans;
 
 	i = 0;
 	sign = 1;
